+++ examples/connectMongoDB.js
@@ -1,0 +1,30 @@
+"use strict";
+/**
+ * MongoDBに接続するサンプル
+ *
+ * @ignore
+ */
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
+Object.defineProperty(exports, "__esModule", { value: true });
+// tslint:disable:no-console
+const sskts = require("../lib/index");
+function main() {
+    return __awaiter(this, void 0, void 0, function* () {
+        const connection = sskts.mongoose.createConnection(process.env.MONGOLAB_URI);
+        console.log('databaseName is', connection.db.databaseName);
+        sskts.mongoose.disconnect();
+    });
+}
+main().then(() => {
+    console.log('success!');
+}).catch((error) => {
+    console.error(error);
+    process.exit(1);
+});